Fixed some warnings Some more code cleanup

+++ src/camera.cpp
@@ -2,8 +2,13 @@
 #include "sample.h"
 #include "ray.h"
 
-Camera::Camera(const float fov, const glm::vec3& pos, const glm::vec3& lookAt, const glm::vec3& up, const unsigned width, const unsigned height) :
-    mPos(pos), mLookAt(lookAt), mUp(up), mWidth(width), mHeight(height), mFov((fov / 180.f) * M_PI)
+Camera::Camera(const float fov, const glm::vec3& pos, const glm::vec3& lookAt, const glm::vec3& up, const unsigned width, const unsigned height)
+    : mPos(pos)
+    , mLookAt(lookAt)
+    , mUp(up)
+    , mWidth(width)
+    , mHeight(height)
+    , mFov((fov / 180.f) * static_cast<float>(M_PI))
 {
     float fovX = 2.f * atanf(tanf(mFov/2.f)*(width / static_cast<float>(height)));
     mAlpha = tanf(fovX/2.f);
